Rename networks.js to networks.js. Rename flag and var to singular. Extract logic from init.js to network.js
